refactor: fix typo in discuss dir name + minor fix PrLPage
