+++ src/pages/problem_list_page/ProblemsListPage.jsx
@@ -35,18 +35,12 @@ const ProblemsListPage = observer(() => {
             }
 
             const data = await response.json();
-            const tasks = data.tasks.map(task => ({
-                id: task._id,
-                name: task.title,
-                description: task.description,
-                difficulty: task.difficulty,
-                likes: task.likes,
-                dislikes: task.dislikes,
-                createdAt: task.created_at,
-                testCases: task.test_cases
-            }));
-            store.setTask(tasks);
-            console.log(tasks);
+            const getTask = [];
+            for (let i = 0; i < data.total; i++){
+                getTask.push({ id: data.tasks[i]._id, name: data.tasks[i].title, description: data.tasks[i].text, difficulty: data.tasks[i].difficulty, likes: data.tasks[i].likes, dislikes: data.tasks[i].dislikes,  createdAt: data.tasks[i].created_at, testCases: data.tasks[i].test_cases});
+            }
+            store.setTask(getTask);
+            console.log(getTask)
         } catch (error) {
             console.error('Error fetching tasks:', error.message || 'Error fetching tasks');
         } finally {
